# optimize ReadAheadInputStream and DocumentUploadService