--- a/sdm/src/main/java/com/sap/cds/sdm/service/DocumentUploadService.java
+++ b/sdm/src/main/java/com/sap/cds/sdm/service/DocumentUploadService.java
@@ -87,7 +87,7 @@ public class DocumentUploadService {
                 long totalSize = cmisDocument.getContentLength();
                 int chunkSize = SDMConstants.CHUNK_SIZE;
 
-                if (totalSize <= chunkSize) {
+                if (totalSize <= 200 * 1024 * 1024) {
                   //  Upload directly if file is ≤ 100MB
                   return uploadSingleChunk(cmisDocument, headers, sdmUrl);
                 } else {
@@ -288,7 +288,7 @@ public class DocumentUploadService {
               logger.info("bytesRead is " + bytesRead);
               // Step 4: Fetch remaining bytes before checking EOF
               long remainingBytes = chunkedStream.getRemainingBytes();
-              logger.info("remainingBytes is " + remainingBytes);
+              logger.debug("remainingBytes is " + remainingBytes);
 
               // Step 5: Check if it's the last chunk
               boolean isLastChunk = bytesRead < chunkSize || chunkedStream.isEOFReached();