--- a/sdm/src/main/java/com/sap/cds/sdm/service/DocumentUploadService.java
+++ b/sdm/src/main/java/com/sap/cds/sdm/service/DocumentUploadService.java
@@ -88,7 +88,7 @@ public class DocumentUploadService {
                 int chunkSize = SDMConstants.CHUNK_SIZE;
 
                 if (totalSize <= 200 * 1024 * 1024) {
-                  //  Upload directly if file is ≤ 100MB
+                  //  Upload directly if file is ≤ 200MB
                   return uploadSingleChunk(cmisDocument, headers, sdmUrl);
                 } else {
                   //  Upload in chunks if file is > 100MB