For handling large file upload in smaller memory there were some optimization required. 
The main reason why with 1G container the 400M file upload was crashing was that the 100M chunk size was higher for the heap to allocate contiguous memorary in that multiple upfront
Made sure we dont use CachedThreadPool instead a fixed size threadpool. Also slightly increased the size of the queue
Also, decided to use 200MB as a criteria for a single chunk upload and changed the chunk size (incase of file size more than 200M) to 20MB. 
Did an extensive testing with 1G and 2G container juggling between 10M and 20M chunk size and confirmed the mem footprint (heap consumption) is stable across. Off course with 10M the used mem of the heap is lesser but even with 20M chunk size the used mem is just linear. 
For now keeping it 20M chunksize because I would not want to emnate twice the network calls (appendContent) from the container to the DI. Rather settling for this trade off 
Here is the detailed test result
<head></head>
2G Container Mem |   |   |   |  
-- | -- | -- | -- | --
  | File size | Upload Time | Heap metrics | Chunk size
  | 272M | 1m 45s | Init: 34 MB, \|Used: 243 MB \|Committed: 319 MB  \|Max: 1297 MB | 20M
  | 562M | 3 min | Init: 34 MB, \|Used: 194 MB \|Committed: 1297 MB \|Max: 1297 MB | 20M
  | 700M | 3 min 41 sec | Init: 34 MB, \|Used: 560 MB \|Committed: 1297 MB \|Max: 1297 MB | 20M
  | 2.12G | 11min 39 sec | Init: 34 MB, \|Used: 199 MB \|Committed: 251 MB  \|Max: 1297 MB | 20M
  | 3.43G | 21 min 5 sec | Init: 34 MB, \|Used: 140 MB \|Committed: 319 MB  \|Max: 1297 MB | 20M
  | 272M | 2 min 5 sec | Init: 34 MB, \|Used: 139 MB \|Committed: 203 MB  \|Max: 1297 MB | 10M
  | 562M | 4min 1 sec | Init: 34 MB, \|Used: 115 MB \|Committed: 203 MB  \|Max: 1297 MB | 10M
  | 700M | 4min 32sec | Init: 34 MB, \|Used: 106 MB \|Committed: 203 MB  \|Max: 1297 MB | 10M
  | 2.13G | 12 min 55sec | Init: 34 MB, \|Used: 115 MB \|Committed: 203 MB  \|Max: 1297 MB | 10M
  | 3.43G | 26 min 35 sec | Init: 34 MB, \|Used: 132 MB \|Committed: 203 MB  \|Max: 1297 MB | 10M
  |   |   |   |  
  |   |   |   |  
  |   |   |   |  
  |   |   |   |  
1G Container Mem |   |   |   |  
  | 272M | 2m 25s | Init: 18 MB, \|Used: 203 MB \|Committed: 238 MB  \|Max: 307 MB | 20M
  | 562M | 5m 36s | Init: 18 MB, \|Used: 211 MB \|Committed: 238 MB  \|Max: 307 MB | 20M
  | 700M | 4m 51s | Init: 18 MB, \|Used: 195 MB \|Committed: 238 MB  \|Max: 307 MB | 20M
  | 2.12G | 13m 51s | Init: 18 MB, \|Used: 152 MB \|Committed: 238 MB  \|Max: 307 MB | 20M
  | 3.43G | 24min 48sec | Init: 18 MB, \|Used: 127 MB \|Committed: 297 MB  \|Max: 307 MB | 20M
  | 272M | 2m 5s | Init: 18 MB, \|Used: 94 MB \|Committed: 175 MB  \|Max: 307 MB | 10M
  | 562M | 3m 39s | Init: 18 MB, \|Used: 164 MB \|Committed: 195 MB  \|Max: 307 MB | 10M
  | 700M | 5m 40s | Init: 18 MB, \|Used: 85 MB \|Committed: 195 MB  \|Max: 307 MB | 10M
  | 2.13G | 17m 3s | Init: 18 MB, \|Used: 82 MB \|Committed: 195 MB  \|Max: 307 MB | 10M
  | 3.43G | 20min 25 sec | Init: 18 MB, \|Used: 110 MB \|Committed: 195 MB  \|Max: 307 MB | 10M
  |   |   |   |  